refactor(mespapiers): Move all ScanResult components in specific folder
